feat: hash tables can now delete clauses for DSR
Hash table lookups for a commutative hash under literals in a
clause allow for O(1) deletions of a clause specified by a
permutation of its literals. This is needed for quick DSR
implementation.

Other fixes include the removal of the sortedness assumption for
deleting a clause in global_data.c, a new parse_clause() function
in sr_parser.c, and some other minor fixes.