feat: ranges for backwards checking
Still needs debugging, but the core
of the implementation is present.

The idea is to store the min/max clause to check
when adding clauses and WPs, and then to use those
ranges during backwards checking. This avoids having
to "push back" the last clause for each literal by
scanning across the formula, which can be time intensive.