feat: partial assignments are ullongs now
Previously, to determine if a literal was set
by the global partial assignment, the magnitude of its
`alpha[var]` generation timestamp value was compared
against `alpha_generation`, with the sign determining
if the assignment is negated or not.

Now, the least-significant bit determines the sign,
and the timestamp values are all `ullong`s.
This gives an ever-so-slight speedup, because
instead of computing an absolute value and
then multiplying by -1 to negate, simple bit
tricks are all that's necessary.

(This speedup really is slight: 0.3% at best.)