feat: refactor into lsr-check source directory
The management of core data structures and helper functions
grew to be too much for one file, so we moved them into
the new `/src/lsr-check/` directory.

(This move was also motivated by the addition of more
verbose error messages for core SR proof checking errors.
In order to avoid having these error messages in the same
file as the core checker logic, we isolated them to their
own file, `lsr-err.c`.)

Error messages updated with both line number in the file,
and line ID as claimed by the clause/deletion line.