feat: optionally ignore deletion of implied units
Addresses https://github.com/ccodel/dsr-trim/issues/3.

Also, more intelligently track which literals are assumed
when assuming the negation of the mapped RAT clause under
the substitution.

Before this commit, `dsr-trim` ran into problems when the
provided DSR proof deleted a clause that, during proof checking,
was actually an implied unit clause, i.e., a clause that became
unit through some round of global unit propagation. We almost
always want to keep implied units, so this commit adds the default
behavior of ignoring the deletion of these clauses (and throw
an error if the clause-to-be-deleted is a true unit).

The implemented solution in this commit is the following:
For each deletion during forwards checking, scan the list of unit
clauses to see if the clause is an implied unit. If it is,
unassign it as a unit clause, then remove any implied unit clauses
derived after this clause, but keep true units. Then re-run UP.

On backwards checking, the solution is to only soft-delete
clauses if they are not currently an implied unit clause.
When watch pointers get restored, each restored clause is checked
to see if it was soft deleted. If it wasn't, don't restore its
watch pointers.