feat(dsr-trim): improved unit literal structures
Previously, unit literals and clauses were treated
with equal weight when it came to computing the
minimal set of UP hints needed per LSR proof line.
However, the previous version had a decision choice
whereby clause "depdency timestamps" were indexed by
clause, even though these timestamps were only updated
for unit clauses, leading to extra array accesses
and an O(#clauses) array, rather than O(#vars).

In this change, the ground truth is the unit literals,
and variables track their index into the `unit_literals`
array so indexes can be found in O(1) time. This change
removed a few for-loops that were required to find the
(minimum) index of a unit clause.

Benchmarking on php-100 and tph20 show a slight slowdown,
but there is likely to be improvement on DRAT proofs
and formulas with a large number of units.

TODOs include tracking the shortest UP derivation chain
for deriving any single unit literal. This can more easily
be accomplished now with indexing, and possibly an additional
array that stores the number of hints needed to derive that literal.